-
Imported-from: https://subversion.assembla.com/svn/AvP/branches/workbranch/trunk@146

+++ 3dc/avp/mempool.c
@@ -11,9 +11,9 @@ static char* MemoryBlocks[MAX_NUM_MEMORY_BLOCK];
 static int CurrentMemoryBlock =-1;
 
 static char* MemoryPoolPtr = 0;
-static unsigned int MemoryLeft = 0;
+static /*unsigned int*/size_t MemoryLeft = 0;
 
-void* PoolAllocateMem(unsigned int amount)
+void* PoolAllocateMem(/*unsigned int*/size_t amount)
 {
 	char* retval;
 //	char buf[100];
